Change int input to float

+++ Exercise5_2_Kornrawit_S.py
@@ -1,4 +1,4 @@
 print("Velocity Calculator Program")
-range = int(input("Please input range in kilometer :"))
-time = int(input("Please input time in hour:"))
+range = float(input("Please input range in kilometer :"))
+time = float(input("Please input time in hour:"))
 print("Your speed =",range/time,"km/h")
